pyproject.toml
[tool.poetry]
name = "eleicao2022"
version = "0.1.0"
description = ""
authors = ["Aluizio Cidral Júnior <cidraljunior@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asn1tools = "^0.164.0"
py7zr = "^0.20.2"
requests = "^2.28.1"
python-slugify = "^6.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
